# botschaft-py/example/openapi.json
{
    "openapi": "3.0.2",
    "info": {
        "title": "FastAPI",
        "version": "0.1.0"
    },
    "paths": {
        "/slack": {
            "get": {
                "tags": [
                    "slack"
                ],
                "summary": "Slack Get",
                "operationId": "slack_get_slack_get",
                "parameters": [
                    {
                        "required": true,
                        "schema": {
                            "title": "Channel",
                            "type": "string"
                        },
                        "name": "channel",
                        "in": "query"
                    },
                    {
                        "required": false,
                        "schema": {
                            "title": "Message",
                            "type": "string"
                        },
                        "name": "message",
                        "in": "query"
                    },
                    {
                        "required": false,
                        "schema": {
                            "title": "Base64 Message",
                            "type": "string"
                        },
                        "name": "base64_message",
                        "in": "query"
                    },
                    {
                        "required": false,
                        "schema": {
                            "title": "Authorization",
                            "type": "string"
                        },
                        "name": "authorization",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "slack"
                ],
                "summary": "Slack Post",
                "operationId": "slack_post_slack_post",
                "parameters": [
                    {
                        "required": false,
                        "schema": {
                            "title": "Authorization",
                            "type": "string"
                        },
                        "name": "authorization",
                        "in": "header"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SlackMessageRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/discord": {
            "get": {
                "tags": [
                    "discord"
                ],
                "summary": "Discord Get",
                "operationId": "discord_get_discord_get",
                "parameters": [
                    {
                        "required": true,
                        "schema": {
                            "title": "Channel",
                            "type": "string"
                        },
                        "name": "channel",
                        "in": "query"
                    },
                    {
                        "required": false,
                        "schema": {
                            "title": "Message",
                            "type": "string"
                        },
                        "name": "message",
                        "in": "query"
                    },
                    {
                        "required": false,
                        "schema": {
                            "title": "Base64 Message",
                            "type": "string"
                        },
                        "name": "base64_message",
                        "in": "query"
                    },
                    {
                        "required": false,
                        "schema": {
                            "title": "Authorization",
                            "type": "string"
                        },
                        "name": "authorization",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "discord"
                ],
                "summary": "Discord Post",
                "operationId": "discord_post_discord_post",
                "parameters": [
                    {
                        "required": false,
                        "schema": {
                            "title": "Authorization",
                            "type": "string"
                        },
                        "name": "authorization",
                        "in": "header"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DiscordMessageRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/twilio": {
            "get": {
                "tags": [
                    "twilio"
                ],
                "summary": "Twilio Message Get",
                "operationId": "twilio_message_get_twilio_get",
                "parameters": [
                    {
                        "required": true,
                        "schema": {
                            "title": "To",
                            "type": "string"
                        },
                        "name": "to",
                        "in": "query"
                    },
                    {
                        "required": false,
                        "schema": {
                            "title": "Message",
                            "type": "string"
                        },
                        "name": "message",
                        "in": "query"
                    },
                    {
                        "required": false,
                        "schema": {
                            "title": "Base64 Message",
                            "type": "string"
                        },
                        "name": "base64_message",
                        "in": "query"
                    },
                    {
                        "required": false,
                        "schema": {
                            "title": "Authorization",
                            "type": "string"
                        },
                        "name": "authorization",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "twilio"
                ],
                "summary": "Twilio Message Post",
                "operationId": "twilio_message_post_twilio_post",
                "parameters": [
                    {
                        "required": false,
                        "schema": {
                            "title": "Authorization",
                            "type": "string"
                        },
                        "name": "authorization",
                        "in": "header"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TwilioMessageRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/sns": {
            "get": {
                "tags": [
                    "sns"
                ],
                "summary": "Sns Get",
                "operationId": "sns_get_sns_get",
                "parameters": [
                    {
                        "required": false,
                        "schema": {
                            "title": "Message",
                            "type": "string"
                        },
                        "name": "message",
                        "in": "query"
                    },
                    {
                        "required": false,
                        "schema": {
                            "title": "Base64 Message",
                            "type": "string"
                        },
                        "name": "base64_message",
                        "in": "query"
                    },
                    {
                        "required": false,
                        "schema": {
                            "title": "Authorization",
                            "type": "string"
                        },
                        "name": "authorization",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "sns"
                ],
                "summary": "Sns Post",
                "operationId": "sns_post_sns_post",
                "parameters": [
                    {
                        "required": false,
                        "schema": {
                            "title": "Authorization",
                            "type": "string"
                        },
                        "name": "authorization",
                        "in": "header"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SnsMessageRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/topic/{topic_name}": {
            "get": {
                "summary": "Topic",
                "operationId": "topic_topic__topic_name__get",
                "parameters": [
                    {
                        "required": true,
                        "schema": {
                            "title": "Topic Name",
                            "type": "string"
                        },
                        "name": "topic_name",
                        "in": "path"
                    },
                    {
                        "required": false,
                        "schema": {
                            "title": "Message",
                            "type": "string"
                        },
                        "name": "message",
                        "in": "query"
                    },
                    {
                        "required": false,
                        "schema": {
                            "title": "Base64 Message",
                            "type": "string"
                        },
                        "name": "base64_message",
                        "in": "query"
                    },
                    {
                        "required": false,
                        "schema": {
                            "title": "Authorization",
                            "type": "string"
                        },
                        "name": "authorization",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/config": {
            "get": {
                "summary": "Config",
                "operationId": "config_config_get",
                "parameters": [
                    {
                        "required": false,
                        "schema": {
                            "title": "Authorization",
                            "type": "string"
                        },
                        "name": "authorization",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Config"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Config": {
                "title": "Config",
                "required": [
                    "providers",
                    "topics"
                ],
                "type": "object",
                "properties": {
                    "providers": {
                        "title": "Providers",
                        "type": "object"
                    },
                    "topics": {
                        "title": "Topics",
                        "type": "array",
                        "items": {}
                    }
                }
            },
            "DiscordMessageRequest": {
                "title": "DiscordMessageRequest",
                "required": [
                    "channel"
                ],
                "type": "object",
                "properties": {
                    "message": {
                        "title": "Message",
                        "type": "string"
                    },
                    "base64_message": {
                        "title": "Base64 Message",
                        "type": "string"
                    },
                    "channel": {
                        "title": "Channel",
                        "type": "string"
                    }
                }
            },
            "HTTPValidationError": {
                "title": "HTTPValidationError",
                "type": "object",
                "properties": {
                    "detail": {
                        "title": "Detail",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        }
                    }
                }
            },
            "SlackMessageRequest": {
                "title": "SlackMessageRequest",
                "required": [
                    "channel"
                ],
                "type": "object",
                "properties": {
                    "message": {
                        "title": "Message",
                        "type": "string"
                    },
                    "base64_message": {
                        "title": "Base64 Message",
                        "type": "string"
                    },
                    "channel": {
                        "title": "Channel",
                        "type": "string"
                    }
                }
            },
            "SnsMessageRequest": {
                "title": "SnsMessageRequest",
                "type": "object",
                "properties": {
                    "message": {
                        "title": "Message",
                        "type": "string"
                    },
                    "base64_message": {
                        "title": "Base64 Message",
                        "type": "string"
                    }
                }
            },
            "TwilioMessageRequest": {
                "title": "TwilioMessageRequest",
                "required": [
                    "to"
                ],
                "type": "object",
                "properties": {
                    "message": {
                        "title": "Message",
                        "type": "string"
                    },
                    "base64_message": {
                        "title": "Base64 Message",
                        "type": "string"
                    },
                    "to": {
                        "title": "To",
                        "type": "string"
                    }
                }
            },
            "ValidationError": {
                "title": "ValidationError",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "type": "object",
                "properties": {
                    "loc": {
                        "title": "Location",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "msg": {
                        "title": "Message",
                        "type": "string"
                    },
                    "type": {
                        "title": "Error Type",
                        "type": "string"
                    }
                }
            }
        }
    }
}